+++ sampleapplication/src/androidTest/java/com/byoutline/sampleapplication/AndroidLibSubmoduleDaggerActivityShould.kt
@@ -1,18 +1,19 @@
 package com.byoutline.sampleapplication
 
-import br.com.concretesolutions.kappuccino.actions.ClickActions.click
 import br.com.concretesolutions.kappuccino.assertions.VisibilityAssertions.displayed
 import com.byoutline.espressohelpers.ActivityRobot
 import com.byoutline.espressohelpers.activityTestRule
 import com.byoutline.espressohelpers.withRobot
 import com.byoutline.sampleapplication.daggerexample.AndroidLibSubmoduleDaggerActivity
+import com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn
 import org.hamcrest.Description
 import org.hamcrest.TypeSafeMatcher
 import org.junit.Rule
 import org.junit.Test
 
 class AndroidLibSubmoduleDaggerActivityShould {
-    @get:Rule val rule = activityTestRule(AndroidLibSubmoduleDaggerActivity::class, AndroidLibSubmoduleDaggerActivityRobot())
+    @get:Rule
+    val rule = activityTestRule(AndroidLibSubmoduleDaggerActivity::class, AndroidLibSubmoduleDaggerActivityRobot())
 
     @Test
     fun displayFragmentAfterBtnClick() {
@@ -23,9 +24,9 @@ class AndroidLibSubmoduleDaggerActivityShould {
     }
 }
 
-class AndroidLibSubmoduleDaggerActivityRobot: ActivityRobot() {
+class AndroidLibSubmoduleDaggerActivityRobot : ActivityRobot() {
     fun openFragment() {
-        click { id(R.id.dagger_example_show_fragment_btn) }
+        clickOn(R.id.dagger_example_show_fragment_btn)
     }
 
     fun verifyThatTextOfFragmentIsDisplayed() {